docs: Add Swagger Operation
