+++ api/src/main/kotlin/com/example/api/member/in/MemberController.kt
@@ -1,13 +1,10 @@
 package com.example.api.member.`in`
 
-import com.example.api.member.`in`.reponse.SuccessAuthenticationResponse
-import com.example.api.member.`in`.request.SendAuthenticationNumberRequest
 import com.example.api.member.`in`.request.SignUpFromEmailRequest
-import com.example.api.member.`in`.request.VerifyAuthenticationNumberRequest
 import com.example.api.member.`in`.request.toCommand
 import com.example.api.response.ApiResponse
-import com.example.core.user.application.`in`.EmailVerifyUseCase
 import com.example.core.user.member.application.`in`.SignUpMemberUseCase
+import io.swagger.v3.oas.annotations.Operation
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
@@ -16,29 +13,13 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/api/v1/member")
 class MemberController(
-    private val emailVerifyUseCase: EmailVerifyUseCase,
     private val signUpMemberUseCase: SignUpMemberUseCase,
 ) {
-    @PostMapping("/email-verify")
-    fun authenticateEmail(
-        @RequestBody request: VerifyAuthenticationNumberRequest,
-    ): ApiResponse<SuccessAuthenticationResponse> {
-        val success = emailVerifyUseCase.verifyEmail(request.toCommand())
-        val response = SuccessAuthenticationResponse(success)
-        return ApiResponse(data = response)
-    }
-
-    @PostMapping("/send/email-verify")
-    fun sendAuthenticationNumberEmail(
-        @RequestBody
-        request: SendAuthenticationNumberRequest,
-    ): ApiResponse<Unit> {
-        emailVerifyUseCase.sendVerifyEmail(request.toCommand())
-        return ApiResponse(data = Unit)
-    }
-
-    @PostMapping("/sign-up/free")
-    fun signUpFreeUser(
+    @PostMapping("/sign-up/email")
+    @Operation(
+        summary = "email 을 통한 회원가입을 합니다.",
+    )
+    fun signUpMemberWithEmail(
         @RequestBody
         request: SignUpFromEmailRequest,
     ): ApiResponse<Unit> {
