+++ api/src/main/kotlin/com/example/api/trainer/in/TrainerController.kt
@@ -1,13 +1,10 @@
 package com.example.api.trainer.`in`
 
-import com.example.api.member.`in`.reponse.SuccessAuthenticationResponse
-import com.example.api.member.`in`.request.SendAuthenticationNumberRequest
-import com.example.api.member.`in`.request.VerifyAuthenticationNumberRequest
 import com.example.api.member.`in`.request.toCommand
 import com.example.api.response.ApiResponse
 import com.example.api.trainer.`in`.request.SignUpTrainerFromEmailRequest
-import com.example.core.user.application.`in`.EmailVerifyUseCase
 import com.example.core.user.trainer.application.`in`.SignUpTrainerUseCase
+import io.swagger.v3.oas.annotations.Operation
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
@@ -16,29 +13,13 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/api/v1/trainer")
 class TrainerController(
-    private val emailVerifyUseCase: EmailVerifyUseCase,
     private val trainerUseCase: SignUpTrainerUseCase,
 ) {
-    @PostMapping("/email-verify")
-    fun authenticateEmail(
-        @RequestBody request: VerifyAuthenticationNumberRequest,
-    ): ApiResponse<SuccessAuthenticationResponse> {
-        val success = emailVerifyUseCase.verifyEmail(request.toCommand())
-        val response = SuccessAuthenticationResponse(success)
-        return ApiResponse(data = response)
-    }
-
-    @PostMapping("/send/email-verify")
-    fun sendAuthenticationNumberEmail(
-        @RequestBody
-        request: SendAuthenticationNumberRequest,
-    ): ApiResponse<Unit> {
-        emailVerifyUseCase.sendVerifyEmail(request.toCommand())
-        return ApiResponse(data = Unit)
-    }
-
-    @PostMapping("/sign-up/trainer")
-    fun signUpTrainer(
+    @PostMapping("/sign-up/email")
+    @Operation(
+        summary = "email 을 사용한 회원가입",
+    )
+    fun signUpTrainerWithEmail(
         @RequestBody
         request: SignUpTrainerFromEmailRequest,
     ): ApiResponse<Unit> {
